+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **BREAKING:** Support max amount quotes ([#7562](https://github.com/MetaMask/core/pull/7562))
+  - Add `TransactionPayController:setIsMaxAmount` messenger action.
+  - Add `isMaxAmount` property to `TransactionData` type.
+  - Add `targetAmount` property to `TransactionPayQuote` and `TransactionPayTotals`.
+  - Update Relay quote requests to use `EXACT_INPUT` trade type when max amount is selected.
+  - Update totals calculation to account for max amount selection.
+
 ### Changed
 
 - Bump `@metamask/controller-utils` from `^11.17.0` to `^11.18.0` ([#7583](https://github.com/MetaMask/core/pull/7583))
 - Bump `@metamask/network-controller` from `^27.1.0` to `^27.2.0` ([#7583](https://github.com/MetaMask/core/pull/7583))
 - Bump `@metamask/assets-controllers` from `^94.0.0` to `^95.0.0` ([#7584](https://github.com/MetaMask/core/pull/7584))
+- Bump `@metamask/transaction-controller` from `^62.7.0` to `^62.8.0` ([#7596](https://github.com/MetaMask/core/pull/7596))
+- Bump `@metamask/bridge-controller` from `^64.3.0` to `^64.4.0` ([#7596](https://github.com/MetaMask/core/pull/7596))
+- Bump `@metamask/bridge-status-controller` from `^64.3.0` to `^64.4.0` ([#7596](https://github.com/MetaMask/core/pull/7596))
 
 ## [10.6.0]